Extracted PathPart into separate file
